Adding Binary Search
Adding Binary Search.
This program asks the user to input the number of elements and then the value of those elements. All the user entered numbers are stored in an array. The program then asks the user to input the search element (the number that needs to be searched in the list). The program searches the entered number in the list using binary search.